# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Micocards API
  version: 0.1.0
  summary: Backend API for the Micocards flashcard service.
  description: |
    REST + JSON. Auth via Authorization: Bearer <jwt>. Non-idempotent writes accept
    an Idempotency-Key header. Status codes follow ADR 0006.
  contact:
    name: Micocards backend
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://v811467.hosted-by-vdsina.com
    description: Production
...
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      operationId: registerUser
      summary: Register a new user.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterUserRequest" }
      responses:
        "201":
          description: registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/auth/login:
    post:
      tags: [auth]
      operationId: loginUser
      summary: Exchange email + password for tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginUserRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshAccessToken
      summary: Rotate the refresh token and mint a new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: rotated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/auth/logout:
    post:
      tags: [auth]
      operationId: logoutUser
      summary: Revoke the supplied refresh-token family.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/me:
    get:
      tags: [me]
      operationId: getCurrentUser
      summary: Fetch the authenticated user's profile.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
    patch:
      tags: [me]
      operationId: updateProfile
      summary: Update the authenticated user's profile.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/me/password:
    post:
      tags: [me]
      operationId: changePassword
      summary: Change the authenticated user's password.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /api/decks/{deckID}/progress:
    parameters:
      - $ref: "#/components/parameters/DeckID"
    get:
      tags: [progress]
      operationId: getUserDeckProgress
      summary: Read the per-card progress for a deck.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserDeckProgress" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /api/practice/sessions/{sessionID}/ratings:
    parameters:
      - $ref: "#/components/parameters/SessionID"
    post:
      tags: [practice]
      operationId: rateCard
      summary: Rate a card inside a tracked session.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RateCardRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RatedCard" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, minLength: 1, maxLength: 128 }
      description: Required for non-idempotent writes per ADR 0005.
    DeckID:
      in: path
      name: deckID
      required: true
      schema: { type: string, format: uuid }
...
    SessionID:
      in: path
      name: sessionID
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unprocessable:
      description: unprocessable entity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Internal:
      description: internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    User:
      type: object
      required: [id, email, displayName, avatarRef, registeredAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        avatarRef:
          oneOf:
            - { type: string, enum: ["none"] }
            - { type: string }
        registeredAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, accessTokenExpiresAt, refreshTokenExpiresAt, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
        refreshTokenExpiresAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    RegisterUserRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 128 }
        displayName: { type: string, minLength: 1, maxLength: 64 }
    LoginUserRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    UpdateProfileRequest:
      type: object
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 1, maxLength: 64 }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 8, maxLength: 128 }
...
    CardRating:
      type: integer
      enum: [0, 1, 2]
      description: 0=DontKnow, 1=StillLearning, 2=KnowKnow.
    RateCardRequest:
      type: object
      required: [cardId, rating]
      properties:
        cardId: { type: string, format: uuid }
        rating: { $ref: "#/components/schemas/CardRating" }
    RatedCard:
      type: object
      required: [sessionId, cardId, rating, ratedAt]
      properties:
        sessionId: { type: string, format: uuid }
        cardId: { type: string, format: uuid }
        rating: { $ref: "#/components/schemas/CardRating" }
        ratedAt: { type: string, format: date-time }
...
    CardProgress:
      type: object
      required: [cardId, rating, lastRatedAt]
      properties:
        cardId: { type: string, format: uuid }
        rating: { $ref: "#/components/schemas/CardRating" }
        lastRatedAt: { type: string, format: date-time }
    UserDeckProgress:
      type: object
      required: [deckId, cardProgress]
      properties:
        deckId: { type: string, format: uuid }
        cardProgress:
          type: array
          items: { $ref: "#/components/schemas/CardProgress" }